+++ aniauth/settings.py
@@ -38,10 +38,10 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'accounts',
+    'logintokens',
 ]
 
-AUTH_USER_MODEL = 'accounts.User'
+AUTH_USER_MODEL = 'logintokens.User'
 
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
